pyproject.toml
[tool.poetry]
name = "generate_image_tool"
version = "0.1.0"
description = ""
authors = ["moarshy <marshath@gmail.com>"]
readme = "README.md"
packages = [
    { include = "generate_image_tool", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
requests = "^2.31.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.2"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git", branch = "feature/tool_calls"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
